Move utils to level_zero lib.
