+++ pmlc/rt/level_zero/level_zero_utils.h
@@ -4,7 +4,7 @@
 #include <utility>
 #include <vector>
 
-#include "pmlc/rt/level_zero/utils/utils.h"
+#include "pmlc/rt/level_zero/utils.h"
 
 namespace pmlc::rt::level_zero::lzu {
