+++ pmlc/testing/CMakeLists.txt
@@ -1,3 +1,5 @@
+include(gflags)
+include(googletest)
 pml_cc_library(
   NAME testing
   SRCS gtest_main.cc
